[refactor] エージェントごとにディレクトリを分けて作業しやすくした
