+++ project/agent/evaluator/evaluator.py
@@ -1,0 +1,8 @@
+from project.model import CustomMessageState
+
+
+def agent_evaluator(state: CustomMessageState) -> CustomMessageState:
+    """
+    エージェントを評価する
+    """
+    return state
