+++ project/agent/generator/generator.py
@@ -1,0 +1,8 @@
+from project.model import CustomMessageState
+
+
+def agent_generator(state: CustomMessageState) -> CustomMessageState:
+    """
+    エージェントプロンプトを生成し保存するノード
+    """
+    return state
