+++ project/agent/optimizer/optimizer.py
@@ -1,0 +1,8 @@
+from project.model import CustomMessageState
+
+
+def agent_optimizer(state: CustomMessageState) -> CustomMessageState:
+    """
+    エージェントを最適化し更新するノード
+    """
+    return state
