+++ project/model.py
@@ -1,0 +1,10 @@
+from langchain_core.messages.utils import AnyMessage
+from langgraph.graph import add_messages
+from pydantic import BaseModel
+from typing_extensions import Annotated
+
+
+class CustomMessageState(BaseModel):
+    should_continue: bool = True
+    should_escalate: bool = False
+    messages: Annotated[list[AnyMessage], add_messages] = []
